requirements.txt
langchain==0.0.143
langchain_visualizer==0.0.20
python-dotenv==1.0.0
tqdm==4.65.0
chromadb==0.3.21
google-search-results==2.4.2
jupyter
openai
tiktoken
pytest
pytest-watch
pytest-vcr
cohere